Fix hardware configuration
Using gitignore didn't work out well, as then Nix won't pick it up (as it's not staged). Hence a different approach was to put the hardware configs in different directories and have an /etc/nixos/machine_id/xyz file that specifies the name of a given machine, allowing for distinguishing the configs.